+++ quarkdown-html-pdf/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    kotlin("jvm")
+}
+
+dependencies {
+    testImplementation(kotlin("test"))
+    testImplementation("org.apache.pdfbox:pdfbox:3.0.6")
+    implementation(project(":quarkdown-core"))
+    implementation(project(":quarkdown-html"))
+    implementation(project(":quarkdown-install-layout-navigator"))
+    implementation(project(":quarkdown-interaction"))
+    implementation(project(":quarkdown-server"))
+}
