+++ quarkdown-html/src/main/kotlin/com/quarkdown/rendering/html/extension/HtmlRendererExtension.kt
@@ -5,8 +5,6 @@ import com.quarkdown.core.document.sub.Subdocument
 import com.quarkdown.core.flavor.RendererFactory
 import com.quarkdown.core.rendering.RenderingComponents
 import com.quarkdown.rendering.html.HtmlExportOptions
-import com.quarkdown.rendering.html.pdf.HtmlPdfExportOptions
-import com.quarkdown.rendering.html.pdf.PdfHtmlPostRendererDecorator
 import com.quarkdown.rendering.html.post.HtmlPostRenderer
 import com.quarkdown.rendering.html.post.HtmlSubdocumentPostRenderer
 
@@ -26,21 +24,4 @@ fun RendererFactory.html(
             Subdocument.Root -> HtmlPostRenderer(context, options.resourcesLayout)
             else -> HtmlSubdocumentPostRenderer(context)
         },
-)
-
-/**
- * The HTML-PDF rendering plug-in produces a PDF document from the HTML output of [html].
- * The outcome is 1:1 with what would be displayed in a Chrome browser.
- */
-fun RendererFactory.htmlPdf(
-    context: Context,
-    pdfOptions: HtmlPdfExportOptions,
-    htmlOptions: HtmlExportOptions,
-) = RenderingComponents(
-    nodeRenderer = accept(HtmlRendererFactoryVisitor(context)),
-    postRenderer =
-        PdfHtmlPostRendererDecorator(
-            HtmlPostRenderer(context, htmlOptions.resourcesLayout),
-            options = pdfOptions,
-        ),
 )
