+++ settings.gradle.kts
@@ -14,6 +14,7 @@ rootProject.name = "quarkdown"
 include("quarkdown-core")
 include("quarkdown-native-library-processor")
 include("quarkdown-html")
+include("quarkdown-html-pdf")
 include("quarkdown-markdown")
 include("quarkdown-plaintext")
 include("quarkdown-cli")
